+++ XADResourceFork.m
@@ -52,7 +52,7 @@
 	// Load the map into memory so that traversing its data structures
 	// doesn't cause countless seeks in compressed or encrypted input streams
 	[handle seekToFileOffset:pos+mapoffset];
-	NSData *mapdata=[handle readDataOfLength:maplength];
+	NSData *mapdata=[handle readDataOfLength:(int)maplength];
 	CSHandle *maphandle=[CSMemoryHandle memoryHandleForReadingData:mapdata];
 
 	[resources release];
@@ -152,7 +152,7 @@
 		NSMutableDictionary *resource=[NSMutableDictionary dictionaryWithObjectsAndKeys:
 			key,@"ID",
 			[NSNumber numberWithInt:attrs],@"Attributes",
-			[NSNumber numberWithUnsignedInt:offset],@"DataOffset",
+			[NSNumber numberWithUnsignedInt:(int)offset],@"DataOffset",
 		nil];
 
 		if(nameoffset!=-1) [resource setObject:[NSNumber numberWithInt:nameoffset] forKey:@"NameOffset"];
